+++ plugins/woocommerce-admin/storybook/main.js
@@ -2,6 +2,8 @@ module.exports = {
 	stories: [
 		// WooCommerce Admin / @woocommerce/components components
 		'../packages/components/src/**/stories/*.js',
+		// WooCommerce Admin / @woocommerce/experimental components
+		'../packages/experimental/src/**/stories/*.js',
 		'../client/**/stories/*.js',
 	],
 	addons: [